take a stab at fixing underlying issue that was messing up 9a

+++ src/nine.rs
@@ -1,7 +1,7 @@
 use crate::computer;
 use crate::computer::{Computer, HaltReason};
 
-pub fn nine_a() -> i32 {
+pub fn nine_a() -> i64 {
     let memory = computer::load_program("src/inputs/9.txt");
     let mut computer = Computer::new(memory, vec![1]);
     computer.run(HaltReason::Exit);
